Callisto.jl

Introduction

Callisto.jl is a simulator for bittide, a system architecture for synchronous distributed computing. A detailed description of the bittide mechanism is available in the papers:

S. Lall, C. Cascaval, M. Izzard and T. Spalink. Modeling and Control of bittide Synchronization. Proceedings of the American Control Conference, 2022. arXiv:2109.14111

S. Lall, C. Cascaval, M. Izzard and T. Spalink. Resistance Distance and Control Performance for bittide Synchronization. Proceedings of the European Control Conference, 2022. arXiv:2111.05296

The model and algorithm used by Callisto follows closely that described in the first of these papers. The code is open-source.

Installation

Start Julia, and at the REPL prompt install Callisto as follows.

julia> using Pkg
julia> Pkg.add(url="https://github.com/bittide/Callisto.jl")

Quickstart

Now you can run a simple simulation:

julia> using Callisto
julia> c = CalOpts()
julia> x = callisto(c)

You can view the output:

julia> xc = parse_callisto_logx(c, x...)
julia> using PlotKit
julia> pzip(a) = Point.(zip(a.x, a.y))
julia> save(drawplot(pzip.(xc.freq)), "frequency.png")

This will save a plot of the frequency against time in the file frequency.png in the Julia working directory.

Callisto.Opts module

Callisto.Opts.Link — Type

Link structure

Callisto.Opts.CalOpts — Type

CalOpts structure

Callisto.Opts.CalOpts — Method

CalOpts(;)

Return a CalOpts structure. Keyword arguments are:

graph latency

Callisto.Opts.make_ugn — Method

make_ugn()

n = 1

Callisto.Piecewise module

Callisto.Piecewise.is_increasing — Method

is_increasing(x)

Return true if x[i-1] < x[i] for all i=2...length(x).

Index

Callisto.Opts.CalOpts

Callisto.Opts.CalOpts

Callisto.Opts.Link

Callisto.Opts.make_ugn

Callisto.Piecewise.is_increasing